NEW setattr methods! replace most set-attr* methods.
The 'setattr' name is very temporary.  All common setting of attributes
are to be handled under this name.  Setting globals and defaults is
planned.   Only setting by handle or id is to be outside these methods
scope.  All these methods return self.

NOTE:  I have found a bit of 'hidden' inconsistency in prior code: some
setting of attributes by reference and some by copy.  This multi is to
make setting by copy the consistent default.  Use the ':ref' flag to set
by reference.

This is experimental and may be backed out or the flag's sense be
inverted.